+++ src/main/java/com/skycat/mystical/common/ConfigModel.java
@@ -37,6 +37,7 @@ public class ConfigModel {
     @Nest public ExplosionsInfestConfig explosionsInfest = new ExplosionsInfestConfig();
     @Nest public BoldSlimesConfig boldSlimes = new BoldSlimesConfig();
     @Nest public ChangingArmorHurtsConfig changingArmorHurts = new ChangingArmorHurtsConfig();
+    @Nest public SoundSwapConfig soundSwap = new SoundSwapConfig();
 
     @SectionHeader("Logging") // Note: Logging as ERROR level does not always mean a critical error.
     public LogLevel failedToSetNightTimerLogLevel = LogLevel.WARN;
@@ -379,6 +380,26 @@ public class ConfigModel {
         }
     }
 
+    public static class SoundSwapConfig {
+        public boolean enabled = true;
+        public LogLevel logLevel = LogLevel.OFF;
+        @PredicateConstraint("weightPredicate")
+        public double weight = 1;
+        @PredicateConstraint("chancePredicate")
+        public double chance = 5.0;
+        @PredicateConstraint("positiveNonzeroIntPredicate")
+        public int numberOfSwaps = 50;
+        public static boolean weightPredicate(double value) {
+            return ConfigModel.weightPredicate(value);
+        }
+        public static boolean chancePredicate(double value) {
+            return ConfigModel.chancePredicate(value);
+        }
+        public static boolean positiveNonzeroIntPredicate(int value) {
+            return ConfigModel.positiveNonzeroIntPredicate(value);
+        }
+    }
+
     /**
      * Verify that the chance is valid.
      * Used instead of {@link RangeConstraint} because it doesn't make a slider.
@@ -408,6 +429,9 @@ public class ConfigModel {
 
     public static boolean positiveNonzeroPredicate(double value) {
         return positivePredicate(value) && value != 0;
+    }
+    public static boolean positiveNonzeroIntPredicate(int value) {
+        return value > 0;
     }
 
 }