+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Fixed nighttime events happening twice (actually this time)
+- Added soundSwap spell